# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: 'DEEO: Dynamic Exam Evaluation Online'
  description: Didactical project for Software Engineering 2 UniTN - Nathaniellee group
host: se2-project-nathaniellee.herokuapp.com
basePath: /api/v0.1
schemes:
- https
consumes:
- application/json
produces:
- application/json
paths:
...
  /exercises:
    get:
      summary: Return the list of all exercises
      produces:
        - application/json
      responses:
        200:
          description: OK
          examples:
            'application/json':
                exercises: [180, 164, 245, 354]
        400:
          description: Bad request
        404:
          description: Not found
        500:
          description: Internal server error

    post:
      tags:
      - Exercises
      summary: Add a new Exercise to the system
      operationId: addExercise
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: Exercise
        description: The Exercise to create
        required: true
        schema:
          $ref: '#/definitions/Exercise'
      responses:
        201:
          description: Created
        400:
          description: Bad request
        401:
          description: Unauthorized
        500:
          description: Internal server error
  /exercises/{exerciseId}:
    get:
      summary: Return the Exercise with the specified ID
      parameters:
        - in: path
          name: exerciseId
          required: true
          type: integer
          description: The ID of the Exercise
      responses:
        200:
          description: A Exercise object
          schema:
            $ref: '#/definitions/Exercise'
        400:
          description: The specified Exercise ID is invalid
        401:
          description: Unauthorized
        404:
          description: A Exercise with the specified ID was not found
        500:
          description: Internal server error

    put:
      summary: Update an existing Exercise
      operationId: updateExercise
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: path
        name: exerciseId
        required: true
        type: integer
        description: The ID of the Exercise
      - in: body
        name: Exercise
        description: The Excercise to update
        required: true
        schema:
          $ref: '#/definitions/Exercise'
      responses:
        200:
          description: OK
        400:
          description: Invalid ID supplied
        401:
          description: Unauthorized
        404:
          description: A Exercise with the specified ID was not found
        500:
          description: Internal server error

    delete:
      summary: Delete a Exercise
      operationId: deleteExercise
      produces:
      - application/json
      parameters:
      - name: api_key
        in: header
        required: false
        type: string
      - name: exerciseId
        in: path
        description: Exercise ID to delete
        required: true
        type: integer
        format: int64
      responses:
        200:
          description: OK
        400:
          description: Invalid ID supplied
        401:
          description: Unauthorized
        404:
          description: A Exercise with the specified ID was not found
        500:
          description: Internal server error
...
definitions:
...
  Exercise:
    type: object
    properties:
      id:
        type: integer
        example: 123
        format: int64
      name:
        type: string
        example: Polpetta
      type:
        type: string
        example: Text box
        enum: [Radio button, Text box, Check box, True or False]
      question:
        type: string
        example: "Which is your favourite ... ?"
      options:
        type: array
        items:
          type: string
        example: [red, green, blue]
    required:
      - id
      - name
      - type
      - question
...